+++ pyon/datastore/test/test_id_factory.py
@@ -1,5 +1,5 @@
 from pyon.util.int_test import IonIntegrationTestCase
-from pyon.datastore.couchdb.id_factory import IDFactory, SaltedTimeIDFactory
+from pyon.datastore.id_factory import IDFactory, SaltedTimeIDFactory
 from nose.plugins.attrib import attr
 
 @attr('UNIT', group='datastore')
